Refactorizacion

+++ backend/src/main/java/com/campito/backend/dto/CuentaBancariaDTOResponse.java
@@ -1,10 +1,12 @@
 package com.campito.backend.dto;
 
+import java.math.BigDecimal;
+
 public record CuentaBancariaDTOResponse(
     Long id,
     String nombre,
     String entidadFinanciera,
-    Float saldoActual
+    BigDecimal saldoActual
 ) {
 
 }
