+++ backend/src/main/java/com/campito/backend/dto/CuotaCreditoDTOResponse.java
@@ -1,12 +1,13 @@
 package com.campito.backend.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record CuotaCreditoDTOResponse(
     Long id,
     int numeroCuota,
     LocalDate fechaVencimiento,
-    Float montoCuota,
+    BigDecimal montoCuota,
     boolean pagada,
     Long idCompraCredito,
     Long idResumenAsociado
