+++ backend/src/main/java/com/campito/backend/dto/PagarResumenTarjetaRequest.java
@@ -1,5 +1,6 @@
 package com.campito.backend.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -24,7 +25,7 @@ public record PagarResumenTarjetaRequest(
     @NotNull(message = "El monto no puede ser nulo")
     @DecimalMin(value = "0.009", message = "El monto debe ser mayor a 0")
     @ValidMonto
-    Float monto,
+    BigDecimal monto,
     @NotBlank(message = "El nombre del usuario no puede estar vacío")
     @Size(max = 100, message = "El nombre completo del usuario no puede exceder los 100 caracteres")
     String nombreCompletoAuditoria,
